pyproject.toml
[tool.poetry]
name = "test-flask"
version = "0.1.0"
description = ""
authors = ["aleksey yakovlev <aleksey.yakovlev.i@zyfra.com>"]
readme = "README.md"
packages = [{include = "test_flask"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^4.0.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
requests = "2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
